requirements.txt
appdirs==1.4.3
beautifulsoup4==4.6.0
certifi==2017.7.27.1
chardet==3.0.4
dj-static==0.0.6
Django==1.11.5
django-appconf==1.0.2
django-compressor==2.2
django-embed-video==1.1.2
django-modelcluster==3.1
django-recaptcha==1.3.1
django-secure==1.0.1
django-sqlcipher==0.1.2
django-sslserver==0.20
django-subdomains==2.1.0
django-taggit==0.22.1
django-treebeard==4.1.2
django-widget-tweaks==1.4.1
djangorestframework==3.6.4
elasticsearch==5.4.0
html5lib==0.999999999
idna==2.6
Markdown==2.6.9
olefile==0.44
packaging==16.8
Pillow==4.2.1
pip-autoremove==0.9.0
Pygments==2.2.0
pyparsing==2.2.0
pytz==2017.2
rcssmin==1.0.6
requests==2.18.4
rjsmin==1.0.12
six==1.10.0
static3==0.7.0
Unidecode==0.4.21
urllib3==1.22
wagtail==1.12.1
wagtail-django-recaptcha==0.2
wagtail-embed-videos==0.3.2
wagtailcommonblocks==0.0.2
wagtailmedia==0.1.4
webencodings==0.5.1
Willow==1.0
pysqlcipher==2.6.10
